Ignore a possilbe ./Carthage/Build folder
This ignore here is prevent projects fetching sdl_ios as a submodule
using Carthage from resulting as `-dirty` in the `git diff`, and always
as modified in the `git status`.

This is due to Carthage copying its Build folder in every submodule in
order to make the whole framework building system work.